services/anomaly-detection/requirements.txt
kafka-python==2.0.2
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
python-dotenv==1.0.0
prometheus_client==0.19.0
